requirements.txt
asgiref==3.3.1
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
defusedxml==0.7.0rc2
Django==3.1.7
django-braces==1.14.0
django-cors-headers==3.7.0
django-oauth-toolkit==1.4.0
django-rest-framework-social-oauth2==1.1.0
django-smart-selects==1.5.9
django-templated-mail==1.1.1
django-webpack-loader==0.7.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-social-oauth2==1.0.9
drf-yasg2==1.19.4
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.27.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
gunicorn==20.0.4
httplib2==0.19.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.9
Pillow==8.1.1
protobuf==3.15.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.16.12
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.4.1
stripe==2.56.0
uritemplate==3.0.1
urllib3==1.26.3
whitenoise==5.2.0
